InsertTraing bug fixed.

+++ src/cloud/Apollo.Backend/Apollo.Api/Convertor.cs
@@ -161,7 +161,7 @@ namespace Apollo.Api
             tr.Certificate = dict.ContainsKey("Certificate") ? (List<string>)dict["Certificate"] : new List<string>();
             tr.Prerequisites = dict.ContainsKey("Prerequisites") ? (List<string>)dict["Prerequisites"] : new List<string>();
             tr.ProductUrl = dict.ContainsKey("ProductUrl") ? new Uri((string)dict["ProductUrl"]) : new Uri("about:blank");
-            tr.Price = dict.ContainsKey("Price") ? (decimal)dict["Price"] : 0M;
+            tr.Price = dict.ContainsKey("Price") ? (double)dict["Price"] : 0;
             tr.Loans = dict.ContainsKey("Loans") ? ToEntityList<Loans>(dict["Loans"] as List<ExpandoObject>, ToLoans) : new List<Loans>();
             tr.TrainingProvider = dict.ContainsKey("TrainingProvider") ? ToTrainingProvider(dict["TrainingProvider"] as ExpandoObject) : null;
             // map other properties here
